# src/openapi.json
{
	"openapi": "3.0.3",
	"info": {
		"title": "Library Management System API",
		"version": "1.1.0",
		"description": "REST API for a library management system — books, authors, users, and the borrow/return workflow, with role-based access control (Admin/Librarian/User).\n\nThis document is generated from JSDoc annotations on the route files (`npm run docs:generate`) — do not hand-edit `docs/API.md` or `src/openapi.json`.",
		"license": {
			"name": "ISC"
		}
	},
	"servers": [
		{
			"url": "http://localhost:5080/api",
			"description": "Local development"
		},
		{
			"url": "{baseUrl}/api",
			"description": "Custom",
			"variables": {
				"baseUrl": {
					"default": "http://localhost:5080"
				}
			}
		}
	],
...
	"components": {
		"securitySchemes": {
			"bearerAuth": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT",
				"description": "Obtain a token from `POST /auth/register`, `POST /auth/login`, or `POST /auth/setup-admin`, then send it as `Authorization: Bearer <token>`."
			}
		},
		"schemas": {
			"Error": {
				"type": "object",
				"properties": {
					"success": {
						"type": "boolean",
						"example": false
					},
					"message": {
						"type": "string",
						"example": "What went wrong"
					},
					"error": {
						"type": "string",
						"description": "Detail, omitted in production for 500s."
					}
				}
			},
			"ValidationError": {
				"type": "object",
				"properties": {
					"success": {
						"type": "boolean",
						"example": false
					},
					"message": {
						"type": "string",
						"example": "Validation failed"
					},
					"errors": {
						"type": "array",
						"items": {
							"type": "object",
							"properties": {
								"field": {
									"type": "string",
									"example": "email"
								},
								"message": {
									"type": "string",
									"example": "Invalid email format"
								}
							}
						}
					}
				}
			},
			"Pagination": {
				"type": "object",
				"properties": {
					"current_page": {
						"type": "integer",
						"example": 1
					},
					"total_pages": {
						"type": "integer",
						"example": 5
					},
					"total_items": {
						"type": "integer",
						"example": 47
					},
					"items_per_page": {
						"type": "integer",
						"example": 10
					},
					"has_next": {
						"type": "boolean",
						"example": true
					},
					"has_prev": {
						"type": "boolean",
						"example": false
					}
				}
			},
...
			"Author": {
				"type": "object",
				"properties": {
					"id": {
						"type": "integer",
						"example": 1
					},
					"first_name": {
						"type": "string",
						"example": "George"
					},
					"last_name": {
						"type": "string",
						"example": "Orwell"
					},
					"email": {
						"type": "string",
						"format": "email",
						"example": "orwell@example.com"
					},
					"phone": {
						"type": "string",
						"nullable": true
					},
					"image": {
						"type": "string",
						"nullable": true
					},
					"date_of_birth": {
						"type": "string",
						"format": "date",
						"nullable": true
					},
					"biography": {
						"type": "string",
						"nullable": true
					},
					"created_at": {
						"type": "string",
						"format": "date-time"
					},
					"updated_at": {
						"type": "string",
						"format": "date-time"
					}
				}
			},
...
		},
		"responses": {
			"Unauthorized": {
				"description": "Missing, invalid, or expired token",
				"content": {
					"application/json": {
						"schema": {
							"$ref": "#/components/schemas/Error"
						},
						"example": {
							"success": false,
							"message": "Access denied. No token provided"
						}
					}
				}
			},
			"Forbidden": {
				"description": "Authenticated, but wrong role or not the resource owner",
				"content": {
					"application/json": {
						"schema": {
							"$ref": "#/components/schemas/Error"
						},
						"example": {
							"success": false,
							"message": "Access denied. Required role(s): Admin. Your role: User"
						}
					}
				}
			},
			"NotFound": {
				"description": "Resource not found",
				"content": {
					"application/json": {
						"schema": {
							"$ref": "#/components/schemas/Error"
						}
					}
				}
			},
			"ValidationFailed": {
				"description": "Request body failed schema validation",
				"content": {
					"application/json": {
						"schema": {
							"$ref": "#/components/schemas/ValidationError"
						}
					}
				}
			},
...
		}
	},
	"security": [
		{
			"bearerAuth": []
		}
	],
	"paths": {
...
		"/authors": {
			"get": {
				"tags": ["Authors"],
				"summary": "List authors",
				"description": "Responses are cached for `AUTHORS_CACHE_TTL_SECONDS` (default 30s), keyed by the full query string, and invalidated on any create/update/delete.\n",
				"security": [],
				"parameters": [
					{
						"name": "page",
						"in": "query",
						"schema": {
							"type": "integer",
							"default": 1
						}
					},
					{
						"name": "limit",
						"in": "query",
						"schema": {
							"type": "integer",
							"default": 10
						}
					},
					{
						"name": "search",
						"in": "query",
						"schema": {
							"type": "string"
						},
						"description": "Matches first_name, last_name, or email."
					},
					{
						"name": "sort_by",
						"in": "query",
						"schema": {
							"type": "string",
							"default": "created_at"
						}
					},
					{
						"name": "order",
						"in": "query",
						"schema": {
							"type": "string",
							"enum": ["asc", "desc"],
							"default": "desc"
						}
					}
				],
				"responses": {
					"200": {
						"description": "Paginated author list.",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"success": {
											"type": "boolean",
											"example": true
										},
										"data": {
											"type": "array",
											"items": {
												"$ref": "#/components/schemas/Author"
											}
										},
										"pagination": {
											"$ref": "#/components/schemas/Pagination"
										}
									}
								}
							}
						}
					}
				}
			},
			"post": {
				"tags": ["Authors"],
				"summary": "Create an author",
				"description": "Requires the Admin or Librarian role.",
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"type": "object",
								"required": ["first_name", "last_name", "email"],
								"properties": {
									"first_name": {
										"type": "string",
										"example": "George"
									},
									"last_name": {
										"type": "string",
										"example": "Orwell"
									},
									"email": {
										"type": "string",
										"format": "email",
										"example": "orwell@example.com"
									},
									"date_of_birth": {
										"type": "string",
										"format": "date",
										"nullable": true
									},
									"biography": {
										"type": "string",
										"nullable": true
									},
									"phone": {
										"type": "string",
										"nullable": true
									},
									"image": {
										"type": "string",
										"nullable": true
									}
								}
							}
						}
					}
				},
				"responses": {
					"201": {
						"description": "Author created.",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"success": {
											"type": "boolean",
											"example": true
										},
										"data": {
											"$ref": "#/components/schemas/Author"
										}
									}
								}
							}
						}
					},
					"400": {
						"$ref": "#/components/responses/ValidationFailed"
					},
					"401": {
						"$ref": "#/components/responses/Unauthorized"
					},
					"403": {
						"$ref": "#/components/responses/Forbidden"
					},
					"409": {
						"description": "An author with this email already exists.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Error"
								}
							}
						}
					}
				}
			}
		},
		"/authors/{id}": {
			"get": {
				"tags": ["Authors"],
				"summary": "Get a single author",
				"security": [],
				"parameters": [
					{
						"name": "id",
						"in": "path",
						"required": true,
						"schema": {
							"type": "integer"
						}
					},
					{
						"name": "include_books",
						"in": "query",
						"schema": {
							"type": "boolean",
							"default": false
						},
						"description": "Embed the author's books and a books_count."
					}
				],
				"responses": {
					"200": {
						"description": "The author.",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"success": {
											"type": "boolean",
											"example": true
										},
										"data": {
											"$ref": "#/components/schemas/Author"
										}
									}
								}
							}
						}
					},
					"404": {
						"$ref": "#/components/responses/NotFound"
					}
				}
			},
			"put": {
				"tags": ["Authors"],
				"summary": "Update an author",
				"description": "Requires the Admin or Librarian role. Partial update — every field is optional.",
				"parameters": [
					{
						"name": "id",
						"in": "path",
						"required": true,
						"schema": {
							"type": "integer"
						}
					}
				],
				"requestBody": {
					"content": {
						"application/json": {
							"schema": {
								"type": "object",
								"properties": {
									"first_name": {
										"type": "string"
									},
									"last_name": {
										"type": "string"
									},
									"email": {
										"type": "string",
										"format": "email"
									},
									"date_of_birth": {
										"type": "string",
										"format": "date",
										"nullable": true
									},
									"biography": {
										"type": "string",
										"nullable": true
									},
									"phone": {
										"type": "string",
										"nullable": true
									},
									"image": {
										"type": "string",
										"nullable": true
									}
								}
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": "Author updated.",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"success": {
											"type": "boolean",
											"example": true
										},
										"data": {
											"$ref": "#/components/schemas/Author"
										}
									}
								}
							}
						}
					},
					"400": {
						"$ref": "#/components/responses/ValidationFailed"
					},
					"401": {
						"$ref": "#/components/responses/Unauthorized"
					},
					"403": {
						"$ref": "#/components/responses/Forbidden"
					},
					"404": {
						"$ref": "#/components/responses/NotFound"
					},
					"409": {
						"description": "Another author already has this email.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Error"
								}
							}
						}
					}
				}
			},
			"delete": {
				"tags": ["Authors"],
				"summary": "Delete an author",
				"description": "Requires the Admin or Librarian role. Fails if the author still has books attached.",
				"parameters": [
					{
						"name": "id",
						"in": "path",
						"required": true,
						"schema": {
							"type": "integer"
						}
					}
				],
				"responses": {
					"200": {
						"description": "Author deleted."
					},
					"401": {
						"$ref": "#/components/responses/Unauthorized"
					},
					"403": {
						"$ref": "#/components/responses/Forbidden"
					},
					"404": {
						"$ref": "#/components/responses/NotFound"
					},
					"409": {
						"description": "Author has books attached — remove or reassign them first.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Error"
								}
							}
						}
					}
				}
			}
		},
...
	}
}